[sc-787] - Update title bar text with new sub routes
