+++ apps/course-authoring/src/renderer/pages/editor/index.tsx
@@ -2,29 +2,22 @@ import React, { useEffect } from 'react';
 import style from './styles.module.scss';
 import { NavigationBar } from '../../components/navigationbar';
 import { Route, Routes } from 'react-router-dom';
-import { PageRoutes, PageNavItems, PageChildren } from './routes';
+import { PageNavItems, PageChildren } from './routes';
 import { PageProps } from '../index.types';
 
 export const PageRoute = '/editor/*';
 export const PageName = 'Course Editor';
 
-interface NewRoute {
-  PageName: string;
-  PageRoute: string;
-}
-interface NewRoutes {
-  pages: NewRoute[];
-}
-
 export const PageElement = ({ handleTitleChange }: PageProps) => {
-  const newRoutes: NewRoutes = { pages: [] };
-
-  Object.entries(PageRoutes).map(page => {
-    newRoutes.pages.push({ PageName: page[0], PageRoute: page[1] });
+  /* Create list of routes from the page navigation items to 
+  update the titlebar list of existing routes */
+  const pageRoutes = PageNavItems.map(page => {
+    console.log('hey');
+    return { label: page.label, link: page.link };
   });
 
   useEffect(() => {
-    handleTitleChange(newRoutes);
+    handleTitleChange(pageRoutes);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
