--- apps/course-authoring/src/renderer/pages/editor/index.types.ts
@@ -1,10 +1,0 @@
-export interface PageRoute {
-  [key: string]: string;
-}
-
-export interface PageNavItem {
-  label: string;
-  link: string;
-}
-
-export type PageNav = PageNavItem[];
